+++ internal/codec/interop/scanner.go
@@ -22,6 +22,7 @@ func OldReaderCodecFields(defaultScanner string) []*service.ConfigField {
 		service.NewScannerField(crFieldCodec).
 			Description("The [scanner](/docs/components/scanners/about) by which the stream of bytes consumed will be broken out into individual messages. Scanners are useful for processing large sources of data without holding the entirety of it within memory. For example, the `csv` scanner allows you to process individual CSV rows without loading the entire CSV file in memory at once.").
 			Default(map[string]any{defaultScanner: map[string]any{}}).
+			Version("4.25.0").
 			Optional(),
 	}
 }