+++ internal/tracing/v2/otel.go
@@ -8,7 +8,7 @@ import (
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 
-	"github.com/benthosdev/benthos/v4/internal/message"
+	"github.com/benthosdev/benthos/v4/public/service"
 )
 
 const (
@@ -17,9 +17,8 @@ const (
 
 // GetSpan returns a span attached to a message part. Returns nil if the part
 // doesn't have a span attached.
-func GetSpan(p *message.Part) *Span {
-	ctx := message.GetContext(p)
-	return GetSpanFromContext(ctx)
+func GetSpan(p *service.Message) *Span {
+	return GetSpanFromContext(p.Context())
 }
 
 // GetSpan returns a span within a context. Returns nil if the context doesn't
@@ -31,30 +30,28 @@ func GetSpanFromContext(ctx context.Context) *Span {
 
 // GetActiveSpan returns a span attached to a message part. Returns nil if the
 // part doesn't have a span attached or it is inactive.
-func GetActiveSpan(p *message.Part) *Span {
-	ctx := message.GetContext(p)
-	t := trace.SpanFromContext(ctx)
+func GetActiveSpan(p *service.Message) *Span {
+	t := trace.SpanFromContext(p.Context())
 	if !t.IsRecording() {
 		return nil
 	}
-	return OtelSpan(ctx, t)
+	return OtelSpan(p.Context(), t)
 }
 
 // GetTraceID returns the traceID from a span attached to a message part. Returns a zeroed traceID if the part
 // doesn't have a span attached.
-func GetTraceID(p *message.Part) string {
-	ctx := message.GetContext(p)
-	span := trace.SpanFromContext(ctx)
+func GetTraceID(p *service.Message) string {
+	span := trace.SpanFromContext(p.Context())
 	return span.SpanContext().TraceID().String()
 }
 
 // WithChildSpan takes a message, extracts a span, creates a new child span,
 // and returns a new message with that span embedded. The original message is
 // unchanged.
-func WithChildSpan(prov trace.TracerProvider, operationName string, part *message.Part) (*message.Part, *Span) {
+func WithChildSpan(prov trace.TracerProvider, operationName string, part *service.Message) (*service.Message, *Span) {
 	span := GetActiveSpan(part)
 	if span == nil {
-		ctx, t := prov.Tracer(name).Start(part.GetContext(), operationName)
+		ctx, t := prov.Tracer(name).Start(part.Context(), operationName)
 		span = OtelSpan(ctx, t)
 		part = part.WithContext(ctx)
 	} else {
@@ -68,9 +65,9 @@ func WithChildSpan(prov trace.TracerProvider, operationName string, part *messag
 // WithChildSpans takes a message, extracts spans per message part, creates new
 // child spans, and returns a new message with those spans embedded. The
 // original message is unchanged.
-func WithChildSpans(prov trace.TracerProvider, operationName string, batch message.Batch) (message.Batch, []*Span) {
+func WithChildSpans(prov trace.TracerProvider, operationName string, batch service.MessageBatch) (service.MessageBatch, []*Span) {
 	spans := make([]*Span, 0, len(batch))
-	newParts := make(message.Batch, len(batch))
+	newParts := make(service.MessageBatch, len(batch))
 	for i, part := range batch {
 		if part == nil {
 			continue
@@ -85,25 +82,25 @@ func WithChildSpans(prov trace.TracerProvider, operationName string, batch messa
 // WithSiblingSpans takes a message, extracts spans per message part, creates
 // new sibling spans, and returns a new message with those spans embedded. The
 // original message is unchanged.
-func WithSiblingSpans(prov trace.TracerProvider, operationName string, batch message.Batch) (message.Batch, []*Span) {
+func WithSiblingSpans(prov trace.TracerProvider, operationName string, batch service.MessageBatch) (service.MessageBatch, []*Span) {
 	spans := make([]*Span, 0, len(batch))
-	newParts := make([]*message.Part, batch.Len())
+	newParts := make([]*service.Message, len(batch))
 	for i, part := range batch {
 		if part == nil {
 			continue
 		}
 		otSpan := GetActiveSpan(part)
 		if otSpan == nil {
-			ctx, t := prov.Tracer(name).Start(part.GetContext(), operationName)
+			ctx, t := prov.Tracer(name).Start(part.Context(), operationName)
 			otSpan = OtelSpan(ctx, t)
 		} else {
 			ctx, t := prov.Tracer(name).Start(
-				part.GetContext(), operationName,
+				part.Context(), operationName,
 				trace.WithLinks(trace.LinkFromContext(otSpan.ctx)),
 			)
 			otSpan = OtelSpan(ctx, t)
 		}
-		newParts[i] = message.WithContext(otSpan.ctx, part)
+		newParts[i] = part.WithContext(otSpan.ctx)
 		spans = append(spans, otSpan)
 	}
 	return newParts, spans
@@ -113,7 +110,7 @@ func WithSiblingSpans(prov trace.TracerProvider, operationName string, batch mes
 
 // InitSpans sets up OpenTracing spans on each message part if one does not
 // already exist.
-func InitSpans(prov trace.TracerProvider, operationName string, batch message.Batch) {
+func InitSpans(prov trace.TracerProvider, operationName string, batch service.MessageBatch) {
 	for i, p := range batch {
 		batch[i] = InitSpan(prov, operationName, p)
 	}
@@ -121,17 +118,17 @@ func InitSpans(prov trace.TracerProvider, operationName string, batch message.Ba
 
 // InitSpan sets up an OpenTracing span on a message part if one does not
 // already exist.
-func InitSpan(prov trace.TracerProvider, operationName string, part *message.Part) *message.Part {
+func InitSpan(prov trace.TracerProvider, operationName string, part *service.Message) *service.Message {
 	if GetActiveSpan(part) != nil {
 		return part
 	}
-	ctx, _ := prov.Tracer(name).Start(part.GetContext(), operationName)
-	return message.WithContext(ctx, part)
+	ctx, _ := prov.Tracer(name).Start(part.Context(), operationName)
+	return part.WithContext(ctx)
 }
 
 // InitSpansFromParentTextMap obtains a span parent reference from a text map
 // and creates child spans for each message.
-func InitSpansFromParentTextMap(prov trace.TracerProvider, operationName string, textMapGeneric map[string]any, batch message.Batch) error {
+func InitSpansFromParentTextMap(prov trace.TracerProvider, operationName string, textMapGeneric map[string]any, batch service.MessageBatch) error {
 	c := propagation.MapCarrier{}
 	for k, v := range textMapGeneric {
 		if vStr, ok := v.(string); ok {
@@ -141,15 +138,15 @@ func InitSpansFromParentTextMap(prov trace.TracerProvider, operationName string,
 
 	textProp := otel.GetTextMapPropagator()
 	for i, p := range batch {
-		ctx := textProp.Extract(p.GetContext(), c)
+		ctx := textProp.Extract(p.Context(), c)
 		pCtx, _ := prov.Tracer(name).Start(ctx, operationName)
-		batch[i] = message.WithContext(pCtx, p)
+		batch[i] = p.WithContext(pCtx)
 	}
 	return nil
 }
 
 // FinishSpans calls Finish on all message parts containing a span.
-func FinishSpans(batch message.Batch) {
+func FinishSpans(batch service.MessageBatch) {
 	for _, p := range batch {
 		if span := GetActiveSpan(p); span != nil {
 			span.unwrap().End()