+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Field `address_cache` added to the `socket_server` input.
-- All inputs with a `codec` field now supports a new field `scanner` to replace it. Scanners are more powerful as they are configured in a structured way similar to all other component types rather than via a single string field, for more information [check out the scanners page](https://www.benthos.dev/docs/components/scanners/about).
+- All inputs with a `codec` field now support a new field `scanner` to replace it. Scanners are more powerful as they are configured in a structured way similar to other component types rather than via a single string field, for more information [check out the scanners page](https://www.benthos.dev/docs/components/scanners/about).
 
 ### Changed